Move Structure to structure.rs module
